major berubah

+++ app/Http/Controllers/Admin/WebinarsControllers.php
@@ -21,32 +21,22 @@ class WebinarsControllers extends Controller
      */
     public function index()
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][24])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
-        $search = '';
-        if (request('category')) {
-            // $category = Category::firstWhere('slug', request('category'));
-            // $title = ' iner ' . $category->name;
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][9])) {
+            throw UnauthorizedException::forPermissions($data);
         }
+        $roleuser = Auth::user()->rolesname;
 
-        if (request('user')) {
-            // $user = User::firstWhere('username', request('user'));
-            // $title = ' by ' . $user->username;
-        }
-
-        $webinars = Webinar::latest()->paginate(10)->withQueryString();
+        $webinars = Webinar::latest()->filter(request(['search']))->paginate(10)->withQueryString();
         $taggers = TagsWebinar::all();
         return view('admin.webinars', [
-            'page' => 'Administrator',
+            'page' => $roleuser,
             'webinars' => $webinars,
-            // 'views' => $kinanda,
             'tages' => $taggers
         ]);
     }
@@ -58,17 +48,18 @@ class WebinarsControllers extends Controller
      */
     public function create()
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][25])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][10])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
+        $roleuser = Auth::user()->rolesname;
         return view('admin.create-webinars', [
-            'page' => 'Administrator',
+            'page' => $roleuser,
             'categories' => CategoryWebinar::all(),
             'tages' => TagsWebinar::all()
         ]);
@@ -82,15 +73,15 @@ class WebinarsControllers extends Controller
      */
     public function store(Request $request)
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][25])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][10])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
 
         $validatedData = $request->validate([
             'title' => 'required',
@@ -119,10 +110,10 @@ class WebinarsControllers extends Controller
         if ($request->file('avatar')) {
             $validatedData['avatar'] = $request->file('avatar')->store('post-image');
         }
-        
+
 
         $slug = SlugService::createSlug(Webinar::class, 'slug', $request->title);
-        $validatedData['typewebinar'] = $request->typewebinar; 
+        $validatedData['typewebinar'] = $request->typewebinar;
         $validatedData['slug'] = $slug;
         $validatedData['latitude'] = '1.2';
         $validatedData['longitude'] = '1.2';
@@ -142,15 +133,15 @@ class WebinarsControllers extends Controller
      */
     public function show($id)
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][24])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][9])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
     }
 
     /**
@@ -161,19 +152,20 @@ class WebinarsControllers extends Controller
      */
     public function edit(Webinar $webinar)
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][26])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
-        
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][11])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
+        $roleuser = Auth::user()->rolesname;
+
         return view('admin.edit-webinars', [
             'webinar' => $webinar,
-            'page' => 'Administrator',
+            'page' => $roleuser,
             'categories' => CategoryWebinar::all(),
             'tages' => TagsWebinar::all()
         ]);
@@ -188,16 +180,16 @@ class WebinarsControllers extends Controller
      */
     public function update(Request $request, Webinar $webinar)
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][26])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
-        
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][11])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
+
         // dd($request);
         $validatedData = $request->validate([
             'title' => 'required',
@@ -226,8 +218,8 @@ class WebinarsControllers extends Controller
         if ($request->file('avatar')) {
             $validatedData['avatar'] = $request->file('avatar')->store('post-image');
         }
-     
-        if($request->slug != $webinar->slug ){
+
+        if ($request->slug != $webinar->slug) {
             $validatedData['slug'] = $request->slug;
         }
         $validatedData['typewebinar'] = $request->typewebinar;
@@ -250,15 +242,15 @@ class WebinarsControllers extends Controller
      */
     public function destroy(Webinar $webinar)
     {
-        // $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
-        //     ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
-        //     ->all();
-        // $data = array(
-        //     "name" => $rolePermissions
-        // );
-        // if (empty($data['name'][27])) {
-        //     throw UnauthorizedException::forPermissions($data);
-        // }
+        $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", Auth::user()['roles'])
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+            ->all();
+        $data = array(
+            "name" => $rolePermissions
+        );
+        if (empty($data['name'][12])) {
+            throw UnauthorizedException::forPermissions($data);
+        }
 
         Webinar::destroy($webinar->id);
         return redirect('/administrator/webinars')->with('success', 'New webinars has been added!');
